backend/requirements.txt
fastapi==0.96.0
uvicorn==0.22.0
python-dotenv==1.0.0
sqlalchemy==2.0.21
pydantic==2.4.2